djangoadvanced/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==4.1.0
black==23.3.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.1
daphne==4.0.0
Django==3.2.19
django-celery-beat==2.5.0
django-cors-headers==4.1.0
django-filter==23.2
django-redis==5.3.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.1.1
flake8==6.0.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.1
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
zope.interface==6.0
